// wall-app/src/components/Table.types.ts
// Generated by https://quicktype.io

export interface TableResponse {
    rankings: Rankings;
}

export interface Rankings {
    matchSeries: MatchSeries;
    ranking:     Ranking[];
}

export interface MatchSeries {
    id:               string;
    uuid:             string;
    allSeasonId:      string;
    name:             string;
    shortName:        string;
    type:             string;
    updated:          string;
    structureUpdated: string;
    resultsUpdated:   string;
    season:           Season;
    hierarchy:        Hierarchy;
    fullHierarchy:    FullHierarchy;
    association:      Association;
}

export interface Association {
    name:      string;
    shortName: string;
}

export interface FullHierarchy {
    hierarchy: Hierarchy[];
}

export interface Hierarchy {
    id:             string;
    name:           string;
    shortName:      string;
    hierarchyLevel: string;
    uuid?:          string;
}

export interface Season {
    id:   string;
    uuid: string;
    name: string;
}

export interface Ranking {
    team:                Team;
    place:               string;
    matchesPlayed:       string;
    wins:                string;
    losses:              string;
    points:              string;
    setPoints:           SetPoints;
    setWinScore:         string;
    setLoseScore:        string;
    setPointDifference:  string;
    setQuotient:         string;
    ballPoints:          string;
    ballWinScore:        string;
    ballLoseScore:       string;
    ballPointDifference: string;
    ballQuotient:        string;
    resultTypes:         ResultTypes;
}

export interface ResultTypes {
    matchResult: MatchResult[];
}

export interface MatchResult {
    result: SetPoints;
    count:  string;
}

export enum SetPoints {
    The03 = "0:3",
    The13 = "1:3",
    The23 = "2:3",
    The30 = "3:0",
    The31 = "3:1",
    The32 = "3:2",
}

export interface Team {
    id:        string;
    uuid:      string;
    name:      string;
    shortName: string;
    clubCode:  string;
    club:      Club;
}

export interface Club {
    name:      string;
    shortName: ShortName;
}

export interface ShortName {
}